+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support OpenMP label and task ID views.
 - Add support for nOS-V non-blocking scheduler server events (`VSN` and `VSn`).
+- Add OpenMP simple breakdown view.
 
 ## [1.11.0] - 2024-11-08